// src/app/domain/models/customer.ts
export class CustomerModel {
    nit!: string
    alias?: string;
    name!: string;
    address?: string;
    phone?: string;
    email?: string;
    occupation?: string;
    contact?: string;
    cellphone?: string;
    userCreated?: string;
    userUpdated?: string;
    dateCreated?: Date;
    dateUpdated?: Date;
    active?: number;
}